feat: decouple utils to separate class
