+++ packages/sdk/src/modules/index.ts
@@ -1,2 +1,3 @@
 export { default as AbstractDatasourceBase } from "./AbstractDatasourceBase"
+export { default as DatasourceUtility } from "./DatasourceUtility"
 export { default as JsonRpcDatasource } from "./JsonRpcDatasource"
